updated read me
