--- eval/.gitignore
@@ -1,2 +1,0 @@
-results/*
-!results/summary*
